attest: Add build flag to gate TrouSerS support
The `tspi` build flag now gates support for TPM 1.2 on Linux. This
allows the attestation libraries to be built without having to first
install the TrouSerS daemon/library/headers.

A build flag of `tspi` instead of `no_tspi` was chosen as go tags tend
to be additive, but I can switch if people think having TrouSerS
support by default is a good idea.

I also updated the documentation to explain all the caveats and gotchas
when using Linux+TrouSerS+TPM 1.2

Signed-off-by: Joe Richey <joerichey@google.com>