[Refactoring] Renamed wrappedTPM20.getPrimaryKeyHandle() into persistHandle() to reflect what the method actually does.
Looking at the implementation, it seems that getPrimaryKeyHandle()
merely ensures that the corresponding object is persisted in the TPM.
Also the method's first return value is always the method's argument
pHnd (if no error occured), thus the first return value may be
redundant. And nowhere in the code the second return value, "whether we
generated a new one", is used; thus it can be safely removed.

Hopefuly this refactoring makes it clearer what the method intends
to do.