Allow rsa.PSSSaltLengthEqualsHash as option
This commit adds support to the salt length set to
rsa.PSSSaltLengthEqualsHash when signing with an RSA key. This is
currently supported, but the user must sspecify the length to the digest
size.

Adding rsa.PSSSaltLengthEqualsHash allows using a key generated by the
TPM for TLS client authentication, Go will set this option by default
if an RSA key is used.

Signed-off-by: Mariano Cano <mariano.cano@gmail.com>