TPM.PCRBanks() should ignore empty PCR banks.
Empty PCR banks are allowed by the TPC spec. But they are no use for the library.
If TPM.PCRBanks() were to return an algorithm for which the TPM has zero PCRs, any following attempt to use this algorithm (i.e. in TPM.PCRs() or TPM.AttestPlatform()) would result in an error anyway. It may be better to omit those algorithms in the first place.